+++ src/App.tsx
@@ -1,11 +1,11 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
 import './App.css'
-import SignUpPage from './pages/SignUpPage'
-import { SignUpProfilePage } from './pages/SignUpProfilePage'
-import HobbySelectionPage from './pages/HobbySelectionPage'
-import { SchoolVerifyPage } from './pages/SchoolVerifyPage'
-import SchoolVerificationPage from './pages/SchoolVerificationPage'
-import SignUpLangTestPage from './pages/SignUpLangTestPage'
+import SignUpPage from './pages/signUp/SignUpPage'
+import { SignUpProfilePage } from './pages/signUp/SignUpProfilePage'
+import HobbySelectionPage from './pages/signUp/HobbySelectionPage'
+import { SchoolVerifyPage } from './pages/signUp/SchoolVerifyPage'
+import SchoolVerificationPage from './pages/signUp/SchoolVerificationPage'
+import SignUpLangTestPage from './pages/signUp/SignUpLangTestPage'
 import SearchHobby from './pages/search/SearchHobby'
 import HomePage from './pages/home/HomePage'
 import MeetingDetailPage from './pages/home/MeetingDetailPage'
